migrate repository to TOOLBOX_PATH
before, all scripts in this repository would use whatever
version of bart comes first in the PATH.
Now, it checks for TOOLBOX_PATH and uses whichever bart is
specified there.

+++ 01_02_03_smallfov/brain/0_prep.sh
@@ -1,12 +1,19 @@
-#!/usr/bin/env bash
-source ../../BART.sh
+#!/bin/bash
+set -e
+
+if [ ! -e $TOOLBOX_PATH/bart ] ; then
+	echo "\$TOOLBOX_PATH is not set correctly!" >&2
+	exit 1
+fi
+export PATH=$TOOLBOX_PATH:$PATH
+
 
 #dataset: same as in Uecker_Magn.Reson.Med._2014
 
-$BART fmac data/alias data/pat data/unders
+bart fmac data/alias data/pat data/unders
 
 # create reference
-$BART fft -u -i 7 data/alias data/tmp
-$BART rss 8 data/tmp data/ref
+bart fft -u -i 7 data/alias data/tmp
+bart rss 8 data/tmp data/ref
 rm data/tmp.*
